Split Qwen35 startup policy helpers
