+++ crates/runner/src/main.rs
@@ -19,6 +19,7 @@ mod qwen35_dflash_engine;
 mod qwen35_kv_trace;
 mod qwen35_prefill;
 mod qwen35_runtime;
+mod qwen35_startup;
 mod qwen35_trace;
 mod qwen35_trace_utils;
 mod qwen35_validation;
