+++ crates/runner/src/qwen35_startup.rs
@@ -1,0 +1,182 @@
+use anyhow::Result;
+
+use crate::registry::{Backend, GpuArch, ModelVariant, Qwen35KernelParams};
+use crate::Cli;
+
+pub(crate) struct Qwen35Startup {
+    pub(crate) text_config: qwen35::config::TextConfig,
+    pub(crate) tokenizer: tokenizers::Tokenizer,
+    pub(crate) prompt_ids: Vec<u32>,
+    pub(crate) context_tokens: usize,
+}
+
+pub(crate) struct Qwen35Policy {
+    pub(crate) trace_kv_cache_enabled: bool,
+}
+
+pub(crate) fn load_qwen35_startup(cli: &Cli) -> Result<Qwen35Startup> {
+    let config = qwen35::config::load_config(&cli.model_dir)
+        .map_err(|e| anyhow::anyhow!("loading config.json: {e}"))?;
+    let text_config = config.text_config;
+    eprintln!(
+        "[config] hidden={} layers={} vocab={} heads={} kv_heads={} head_dim={}",
+        text_config.hidden_size,
+        text_config.num_hidden_layers,
+        text_config.vocab_size,
+        text_config.num_attention_heads,
+        text_config.num_key_value_heads,
+        text_config.head_dim,
+    );
+
+    let tokenizer_path = cli.model_dir.join("tokenizer.json");
+    let tokenizer = tokenizers::Tokenizer::from_file(&tokenizer_path)
+        .map_err(|e| anyhow::anyhow!("load tokenizer: {e}"))?;
+    let encoding = tokenizer
+        .encode(cli.prompt.as_str(), !cli.prompt_no_special_tokens)
+        .map_err(|e| anyhow::anyhow!("tokenize: {e}"))?;
+    let prompt_ids: Vec<u32> = encoding.get_ids().to_vec();
+    eprintln!("[tokenizer] prompt_tokens={}", prompt_ids.len());
+    if prompt_ids.is_empty() {
+        anyhow::bail!("empty prompt after tokenization");
+    }
+
+    let context_tokens = cli
+        .context_size
+        .unwrap_or(prompt_ids.len() + cli.max_new_tokens);
+
+    Ok(Qwen35Startup {
+        text_config,
+        tokenizer,
+        prompt_ids,
+        context_tokens,
+    })
+}
+
+pub(crate) fn validate_qwen35_startup(
+    cli: &Cli,
+    model_variant: &ModelVariant,
+    params: &Qwen35KernelParams,
+    backend: Backend,
+    registry_arch: &GpuArch,
+    q4km_like: bool,
+) -> Result<Qwen35Policy> {
+    if cli.trace_prefill_layers && !cli.validate {
+        anyhow::bail!("--trace-prefill-layers requires --validate");
+    }
+    if cli.trace_oracle_prefill_layer.is_some() && !cli.validate {
+        anyhow::bail!("--trace-oracle-prefill-layer requires --validate");
+    }
+    if cli.trace_kv_fp8_cache && !cli.kv_fp8 {
+        anyhow::bail!("--trace-kv-fp8-cache requires --kv-fp8");
+    }
+    let trace_kv_cache_enabled = cli.trace_kv_cache || cli.trace_kv_fp8_cache;
+    if cli.force_kernel_decode && cli.force_component_decode {
+        anyhow::bail!("Choose at most one of --force-kernel-decode or --force-component-decode");
+    }
+    if cli.trace_component_input_layer.is_some() && !cli.force_component_decode {
+        anyhow::bail!("--trace-component-input-layer requires --force-component-decode");
+    }
+    if cli.trace_component_layer.is_some() && !cli.force_component_decode {
+        anyhow::bail!("--trace-component-layer requires --force-component-decode");
+    }
+    if cli.trace_component_linear_layer.is_some() && !cli.force_component_decode {
+        anyhow::bail!("--trace-component-linear-layer requires --force-component-decode");
+    }
+    if cli.trace_component_linear_state_layer.is_some() && !cli.force_component_decode {
+        anyhow::bail!("--trace-component-linear-state-layer requires --force-component-decode");
+    }
+    if cli.trace_persistent_input_layer.is_some()
+        && !(params.use_4b_kernel
+            && !cli.force_component_decode
+            && (cli.batch_size > 1 || cli.force_kernel_decode || cli.kv_fp8))
+    {
+        anyhow::bail!("--trace-persistent-input-layer requires the real 4B persistent kernel path");
+    }
+    if cli.trace_persistent_linear_state_layer.is_some()
+        && !(params.use_4b_kernel
+            && !cli.force_component_decode
+            && (cli.batch_size > 1 || cli.force_kernel_decode || cli.kv_fp8))
+    {
+        anyhow::bail!(
+            "--trace-persistent-linear-state-layer requires the real 4B persistent kernel path"
+        );
+    }
+    if cli.trace_persistent_full_attn_layer.is_some()
+        && !(params.use_4b_kernel
+            && !cli.force_component_decode
+            && (cli.batch_size > 1 || cli.force_kernel_decode))
+    {
+        anyhow::bail!(
+            "--trace-persistent-full-attn-layer requires the real 4B persistent kernel path"
+        );
+    }
+    if cli.trace_persistent_linear_layer.is_some()
+        && !(params.use_4b_kernel
+            && !cli.force_component_decode
+            && (cli.batch_size > 1 || cli.force_kernel_decode)
+            && !cli.kv_fp8)
+    {
+        anyhow::bail!(
+            "--trace-persistent-linear-layer requires the real 4B persistent BF16 kernel path"
+        );
+    }
+
+    if backend == Backend::Cuda {
+        let qwen35_sm86 = matches!(
+            model_variant,
+            ModelVariant::Qwen3_5_0_8B
+                | ModelVariant::Qwen3_5_2B
+                | ModelVariant::Qwen3_5_4B
+                | ModelVariant::Qwen3_5_9B
+        ) && *registry_arch == GpuArch::Sm86;
+        if cli.int4 && !qwen35_sm86 {
+            anyhow::bail!("CUDA --int4 currently supports only Qwen3.5 on sm86");
+        }
+        if cli.fp8_runtime && !(qwen35_sm86 || *model_variant == ModelVariant::Qwen3_6_27B) {
+            anyhow::bail!(
+                "CUDA --fp8-runtime currently supports only Qwen3.5 and qwen3.6-27b on sm86"
+            );
+        }
+        if cli.kv_fp8 {
+            if !qwen35_sm86 {
+                anyhow::bail!("CUDA --kv-fp8 currently supports only Qwen3.5 on sm86");
+            }
+            if std::env::var_os("SUPERSONIC_DEBUG_ENABLE_CUDA_KV_FP8_BF16_SIDECAR").is_none() {
+                std::env::set_var("SUPERSONIC_DEBUG_KV_FP8_BF16_SIDECAR_WINDOW", "128");
+                eprintln!(
+                    "[cuda] KV-FP8 BF16 sidecar window capped to the most recent 128 tokens on CUDA; \
+                     set SUPERSONIC_DEBUG_ENABLE_CUDA_KV_FP8_BF16_SIDECAR=1 \
+                     to restore full-prefix debug sidecar coverage"
+                );
+            }
+        }
+    } else if backend == Backend::Metal {
+        if !matches!(
+            model_variant,
+            ModelVariant::Qwen3_5_0_8B | ModelVariant::Qwen3_5_2B
+        ) {
+            anyhow::bail!("Metal only supports --model qwen3.5-0.8b or qwen3.5-2b");
+        }
+        if q4km_like {
+            anyhow::bail!("Metal does not support --q4km/--q4km-gptq on Qwen3.5 yet");
+        }
+        if cli.fp8_runtime {
+            anyhow::bail!("Metal does not support --fp8-runtime on Qwen3.5 yet");
+        }
+        if cli.kv_fp8 {
+            anyhow::bail!("Metal does not support --kv-fp8 on Qwen3.5 yet");
+        }
+        if cli.batch_size != 1 {
+            anyhow::bail!("Metal only supports --batch-size 1");
+        }
+        if cli.force_kernel_decode || cli.force_component_decode {
+            anyhow::bail!(
+                "Metal does not support --force-kernel-decode or --force-component-decode"
+            );
+        }
+    }
+
+    Ok(Qwen35Policy {
+        trace_kv_cache_enabled,
+    })
+}
